AI_API/requirements.txt
pandas
sqlalchemy
fastapi
pydantic
ipython
textwrap3
google-generativeai
langchain-text-splitter
langchain-google-genai
langchain
json
